Make wallet-context an optional resource
Information found in the wallet context json should not be critical.
Failing to access this resource should not block the app. Instead,
this context should simply provide information that improve user
experience, such as the warning regarding the mempool.

The wallet params are now hardcoded, including the trampoline fees
settings. Note that most of the data in the json are legacy and were
already unused.

The logic fetching the wallet context can be relaxed and simplified.
There's also no need to store this context anymore.