(android) Merge current & archive files when exporting logs
On android we use SLF4j to manage logs, with rolling. When exporting
the logs, we should also export the archive, not only the current
log file. These two files are now merged in a temp log file when
exporting logs.

We also stop storing a 2nd archive file, as it is not used.