(android) Add overview of recently rejected payments in home
It's easy to miss payment rejected notifications. System notifications
may be filtered out, or lost in noise if the user has many unread
notifications. In that case, the user may not notice that payments
were missed, and why.

This is similar to the work done in https://github.com/ACINQ/phoenix/pull/423
for iOS, except that the notification message displays the count of
missed payments instead of providing the details (which takes a lot
of place).

Also, only recent failures are displayed (< 3 hours).