Show swap-in time-outs in wallet screen and notifications
Phoenix uses the swap-in potentiam scheme to trustlessly swap
on-chain deposits to Lightning with 0 confirmations channels.

This scheme introduces a duration of 4 months during which swaps
can be attempted. After this 4 months have passed, deposits that
have not been swapped time out and enter a grace period of two
months. After this two months period has passed, these deposits
can be unilaterally spent by the user's wallet to any address.

This commit improves the swap-in screen by discrimating funds
according to their status (confirming, swappable, timed-out,
and cancelled), based on their confirmation count.

An important notice is also added to the home screen, if a swap
is going to time out soon (< 1 month).

Also, the notifications for pending deposits will mention the
timeout if it is available at the time of the notification.