(ios) Display more accurate lightning fee as a percentage (#449)
The actual fee amount is: 4 sats + 0.4%. So this meant the user 
might see an effective fee of "0.5%" or "0.7%" when sending smaller 
payments.  They might proceed to think this is the fee for all payments.

Now the effective fee displays the base fee + the percentage fee when
needed, or just the percentage fee if the payment is large enough.