(android) Add fallback to seed when keystore fails
Phoenix should gracefully handle errors in the Android KeyStore.
When the key encrypting the seed file on the FS is unavailable
or corrupted in the keystore, the app should let the user restore
their wallet using their seed.

The app also allows the user to export the app logs when such
errors occur, to help with debugging if need be.