(android) Remove plurals resources
Plurals resources cause crashes on Samsung devices for some languages,
probably because of a bug on Samsung side which we cannot fix. Error
is `Resources$NotFoundException` although the resource does exist.
This could be a compat issue with Compose, or a Localization issue.

In any case this may prevent users from opening the wallet. The solution
is for now to remove plurals resources from the app.