Add database and queries for inbound liquidity payments
A liquidity lease rate is provided to the peer at startup
using default values stored in the NodeParamsManager.

Aggregated queries have been updated.

Note that inbound liquidity payments use the `lockedAt`
timestamp to define if the payment is complete or not.

Also added a new `WalletPaymentId` for inbound liquidity
payments, with internal code 6.

Tests are missing. Cloud wrappers for inbound liquidity
payments are not yet implemented.