(android) Add UI for requesting liquidity
A new screen has been added to request liquidity from
the peer using liquidity-ads. The user can pick the an
amount, and get an estimation of the cost (mining +
service fee) of the liquidity. Then accept the offer, or
cancel the request by leaving the screen. The liquidity
is done with a splice.

A button to that screen has been added in the home screen
and in the liquidity policy screen.

The channels view screen also shows the current inbound
liquidity, using a linear progress bar.