(android) Add support for liquidity ads (#489)
This commit allows users to request inbound liquidity
from their peer on the Android app. The iOS app will be 
updated later, in another commit.

* Use lightning-kmp 1.5.13

* Add database and queries for inbound liquidity payments

A liquidity lease rate is provided to the peer at startup
using default values stored in the NodeParamsManager.

Aggregated queries have been updated.

Note that inbound liquidity payments use the `lockedAt`
timestamp to define if the payment is complete or not.

Also added a new `WalletPaymentId` for inbound liquidity
payments, with internal code 6.

* (android) Add UI for requesting liquidity

A new screen has been added to request liquidity from
the peer using liquidity-ads. The user can pick the 
amount, and get an estimation of the cost (mining +
service fee) of the liquidity. Then accept the offer, or
cancel the request by leaving the screen. The liquidity
is done with a splice.

A button to that screen has been added in the home
and the liquidity policy screens.

The channels view screen also shows the current inbound
liquidity, using a linear progress bar.