Upgrade to kotlin 1.9.22
Several libraries have been upgraded:
- ktor to 2.3.7
- kotlinx.coroutines to 1.7.3
- kotlinx.serialization to 1.6.2

tor-mobile-kmp and secp256k1 have been also upgraded to
development version using kotlin 1.9 (we'll use tags
before this commit is merged to master).

Note: the `iosSimulatorArm64` target has been removed for
now as it's not yet supported by tor-mobile-kmp. This will
be reverted once the tor library is fixed.

The kodein-memory library has also been removed. This
library was used to write the logs to disk in the shared
module. However we were using an old version, and we are
already planning to move to another log library and to
delegate writing logs to disk to the application layer.
It's not worth spending time upgrading ; instead this
dependency is removed, and the code should be cleaned up:

-> removed LogMemory
-> removed usage of APIs from this lib (freeze calls)
-> remove most code from the LogConfiguration controller
   which has to be removed (will be done in another commit
   as the UI would be impacted)