(android) Add in-flight payments counter in home
Pending payments can timeout if the user's device is not
background-friendly, and the app is not started for a while.
Timeout trigger force-close, so they need to be avoided as
much as possible. This user should be made aware of that.

A common issue is GrapheneOS devices which can work in the
background, however cannot receive the silent push notifs
from the ACINQ peer, and thus are vulnerable to timeouts.

This commmit is a wip - a background job starting the app
regularly is also needed.