(android) Use taproot-rotated addresses by default (wip)
By default the app should use the last available address
exposed by the swap-in wallet. The user can switch to the
legacy address if need be, for compatibility purposes with
older services.

Previously used addresses can be displayed in the Receive
screen for now.

Since the swap-in wallet may take time to synchronize all
used addresses at startup, the app stores the last known
unused index in the preferences, and uses that index when
the app starts. This way, the user does not have to wait
for an address to be shown before he can deposit funds to
Phoenix. There's a risk of address reuse but it's small
enough and the UX benefits are significant if the wallet
must check a lot of addresses.