Upgrade to Kotlin 1.9 (#509)
Kotlin 1.9.22 brings many improvements, for details see:
https://github.com/ACINQ/lightning-kmp/pull/565

Several dependencies have been upgraded:
- ktor to 2.3.7
- ktor to 0.13.0
- lightning-kmp to 0.16.x
- gradle to 8.0

Note: the `iosSimulatorArm64` target has been removed for
now as it's not yet supported by tor-mobile-kmp. This will
be reverted once the tor library is fixed.

The kodein-memory library has also been removed. This
library was used to write the logs to disk in the shared
module. However we were using an old version, and we are
already planning to move to another log library and to
delegate writing logs to disk to the application layer.
It's not worth spending time upgrading ; instead this
dependency is removed, and the code was cleaned up:

-> removed LogMemory
-> removed usage of APIs from this lib (freeze calls)
-> remove most code from the LogConfiguration controller
   which has to be removed (will be done in another commit
   as the UI would be impacted)

---------

Co-authored-by: Robbie Hanson <304604+robbiehanson@users.noreply.github.com>