Use lightning-kmp with taproot swaps and rotation
The swap-in wallet now exposes a flow containing the current
address, and a flow of addresses that have been used. Those
addresses use taproot and txs should be lighter and more
private.

Note that the wallet does not expose future unused addreses,
only past addresses that have already received funds, or the
most recent (index-wise) address that thas not been used.

The legacy address is still available and will still work,
however usage should be discouraged.

There have been some api changes in a few methods.

The swap-in model have been removed from the ReceiveController
since it's not used anymore, and we want to move away from the
MVI pattern.