(android) Add fee/liquidity disclaimers in several screens
A confirmation message has been added in the request liquidity
screen, to clarify that the added liquidity is consumed by
receiving payment. The duration help has also been amended.

A confirmation message has been added in the splice-out screen,
before sending the payment, to clarify that sending funds on-chain
diminishes mechanically the capacity of the channel.

A warning message has been added to the on-chain receive screen
to remind the user that on-chain deposits do not use available
liquidity, and will incur fees.