Add option to overpay LN payments (#541)
The option to overpay has been added to both iOS and Android, in 
Settings > Payment options.

Overpayment is still capped to 2 times the requested amount.

On Android, +- 5% tipping buttons have been added when sending 
LN payments. The iOS app already had buttons to add tips.

On Android, convert the amount when switching unit, instead of
only changing the unit. It causes some rounding issues, but it's 
a reasonable tradeoff.

---------

Co-authored-by: Robbie Hanson <304604+robbiehanson@users.noreply.github.com>