Add fee/liquidity disclaimers in several screens (#539)
When requesting liquidity, a confirmation message is displayed
with a cancel/confirm button. The message clarifies that the
liquidity will be consumed when receiving payments. The help
message for the duration has also been updated.

A confirmation message has been also added in the splice-out 
screen, to clarify that sending funds on-chain does impact the 
capacity of the channel (but not liquidity).

---------

Co-authored-by: Robbie Hanson <304604+robbiehanson@users.noreply.github.com>