(android) Add UI to attach offer to a contact
A photo can be attached to the contact, though this
version only uses the Preview feature and not a full
size photograph. Also needs to be able to import an
image from the device's gallery.

The dialog to edit/attach a custom description has
also been update to use a bottom sheet layout, for
consistency reasons.