(android) Add custom PIN code support
A 6-digits PIN code can now be defined to restrict access to
the application, in addition to the System screen lock option.
Both can be used at the same time, or just one.

This PIN code is specific to Phoenix and can be different from
the System PIN.

This change is similar to the PIN code update for iOS (#560).