(android) Add button to contacts in scanner view
The back button in scanner view has been moved to the
top of the screen. The visor has been removed and the
scanner target is larger.

When adding a new contact, the offer can be entered
from the modal window (manual input or scanner).

The contacts list can now be filtered on name.