(android) Move receive-offer to a bottom sheet
We only have 2 tabs now, one for LN and one for on-chain, which
is easier to understand. The offer is displayed in a modal window
accessed after clicking on a button below the bolt11 QR.

Wording in the help has been updated as well.

The QR image component has been updated (brightness, save to disk).