(android) Add support for Bolt12 offers (#575)
The offer displayed is the default offer as computed in
NodeParams.defaultOffer. An helper method has been added
in NodeParamsManager to easily get that offer.

Added several screen to attach an offer to a contact, manage
the contacts, and to list contacts. This list of contacts can be 
reached from the scanner view and the settings menu.

Contacts are stored locally in `appdb.sqlite`. A new manager
`ContactsManager` has been added.

Parsing offers is done through the ScanDataController, but
paying it is done with `Peer.payOffer`. A message can be
attached when paying an offer.

Added a new setting in the payment options to use a random 
payer key.