(android) Split app into 2 flavours (google & foss)
The google flavor is able to use FCM notifications for background
processing. A FCM token is registered automatically.

The foss flavor does not embed any google play services and instead
relies on a permanent foreground service to process payments in the
background. This experimental service is opt-in and needs to be
manually started by the user. The user can configure whether or not
a partial wake lock should be acquired (enabled by default).