Make payer-key option specific to each contact
Whether to use a random payer key or the offer's key
is now defined on a per-contact basis. By default,
when adding a new contact, this option is set to use
the offer key. This option can be changed anytime and
is persisted in the database.

Added the UI to view and modify this option in Android.