Accept email-like address with a prefix
Email-like addresses may be shared with a QR code, for example
in a conference. To prevent the scanners from opening the address
in an email application, which would not work, people typically
use a prefix (e.g. `lightning`) so that a Lightning app is used.

Fixes #616

(cherry picked from commit 5a22661288b3eb9c6ccc9ac1ed00380d015b1411)