Update phoenix-shared to new on-the-fly
- Updated the liquidity policy (see NodeParamsManager). We
are using a policy that does not target additional liquidity
and that does not use fee credit yet.

- Removed the code updating the peer's swap feerate (it's now
provided by the LSP)

- Updated the received-with database object with new types:
incoming LN payments may contain a funding fee ; payments may
be received through a fee credit (not enabled for now).

- Updated the liquidity-purchase database objects with new
purchase & payment-details type. The lease data are now
legacy and are removed wherever possible.

Note: for the inbound liquidity db objects, we are moving
away from the type_version pattern for serialisation. Instead
the json data type is embedded inside the json by the kotlinx
serialisation library. This makes the code less verbose.

- Added a liquidity purchase wrapper for cloud data.

- Updated the Notification types with new rejection options.
We try to match in this update the serialisation pattern that
was mentioned above.