(android) Use lightning-kmp method to determine whether liquidity fee is shown or not
See https://github.com/ACINQ/lightning-kmp/pull/706

Methods feePaidFromChannelBalance and feePaidFromFutureHtlc have been added
to check whether an inbound liquidity purchase is paid from balance or future
htlc. This is used to know when to display a liquidity fee or not, and to get
the liquidity fee for an incoming payment.

Also fixed some UI issues with descriptions and improved txid and channel id
buttons in technical screen.