Support new on-the-fly funding (#632)
Now using lightning-kmp 1.8.0.

- Updated the liquidity policy (see NodeParamsManager). We
are using a policy that does not target additional liquidity
and that does not use fee credit yet.

- Updated the received-with database object with new types
- Updated the liquidity-purchase database objects with new
purchase & payment-details types. The lease data are now
legacy and are removed wherever possible.

- Updated the iOS and Android payment screens for liquidity 
purchases. Liquidity events are now stored as separate payments.
These payments may be linked to incoming payments, if so
a button is added in the liquidity screen to navigate to these
payments.

- (android) Completed the Portuguese (BR) translation.

---------

Co-authored-by: Robbie Hanson <304604+robbiehanson@users.noreply.github.com>