Add db method to link liquidity with swap-ins
Swap-ins that create a new channel trigger a liquidity event. To be
consistent, these liquidity events should display a link to the swap
payment. To do this we rely on the existing table that links known
tx ids to wallet-payments (in this case, incoming payments).

(android) added the code to fetch these payments and updated the UI.