Ignore inbound liquidity purchases for swap-ins creating channels
Swap-ins creating new channels trigger a liquidity event for 1 sat
for technical reasons. These purchases should be ignored and not
be visible in the history (the fees are already taken into account
in the swap-in payment).

To do that a new column has been added so we can filter the amount
in the SQL query. This column is nullable to make migration painless
so this field will be null for legacy purchases.