Split liquidity mining fee between local and purchase fee
With https://github.com/ACINQ/lightning-kmp/pull/709 the mining
fee is split between local and purchase fees. However, for compat
with legacy leases, we do not store the local fee separately in
the database. Instead we store the full mining fee and calculate
the local fee by subtracting this full fee with the purchase fee.