Link liquidity events to swap-in payments (#636)
Swap-ins that create a new channel trigger a liquidity event. To be
consistent, these liquidity events should display a link to the swap
payment.

Also lightning-kmp 1.8.1 fixes an issue with swap-in payments mining
fee, which is now split between local and purchase fees. For compat
with legacy leases, we do not store the local fee separately in
the database. Instead we store the full mining fee and retrieve
the local fee by subtracting this full fee with the purchase fee.

---------

Co-authored-by: Robbie Hanson <304604+robbiehanson@users.noreply.github.com>