make tor requirements for custom electrum servers opt-out
When tor is enabled, Phoenix will require electrum servers to use an
onion address. This way, if the Tor VPN fails, connection will not
be established. However, some users may want to opt-out, which is
valid if they are connecting to their own server and don't mind
connecting to it on clearnet.