run db-migration tests on an android environment
Sample database files have been moved to commonTest. They are loaded
using okio, which has been added as a dependency for the commonTest
module.

The migration test now expects a method that sets up the testing
environment, and which is ran before the migration tests. This method
has been implemented on Android (only) and uses robolectric to create
a simulated Android environment, without having to install a real
Android simulator or a device. The sample db files are copied to the
database folder expected by Android, which allows us to use the
regular paymentsdb driver.

The migrated files are copied into the test-results folder for
debugging if needed.