Remove Tor embedded library and use onion addresses (#662)
The Tor option has been changed. Instead of starting an embedded Tor
library when the app starts, it requires Phoenix to use onion addresses
for the Peer and Electrum connections and delegates Tor management
to a third party app.

This is safer, as a failure in the Tor proxy cannot lead to a leak. It also 
improves payments reliability, since the Tor proxy app being a VPN app, 
its connection is much more reliable especially when Phoenix is in the
background. This app is also better maintained, and this frees Phoenix
from the complexity of managing a cross-platform Tor library.

Note that the custom Electrum server setting can opt-out from the onion
address requirement.

---------

Co-authored-by: Robbie Hanson <304604+robbiehanson@users.noreply.github.com>