Remove the phoenix-legacy module (#671)
This commits removes the old 1.x Android legacy code, which was using
the eclair-core Scala library. This version has been deprecated 18 months 
ago and all legacy channels have now been closed.

This includes code that was added in the modern android app to manage
the switch to the legacy app, as well as the migration. Gone as well
is the special init TLV sent to the peer to check for legacy channels.