(android) Add system to manually import/export payments history
The user can now manually export the payments database to a public folder
of the device. The file is zipped and encrypted with ChaCha20Poly1305, and
another option with AES and javax-crypto has been added for comparison.

When restoring a wallet, the user can also pick a payments backup file after
typing down the seed. This is optional.