(android) Hide auto purchase amount when payment is received
The logic is thus: when we receive a payment that triggers an automated
liquidity purchase, that purchase is displayed as a Channel Management
fee. But that fee has already been deducted from the incoming payment's
amount. We should not display that fee twice. So, if we know the payment
has indeed been received (done by checking incomingPaymentReceivedAt),
then we hide the purchase amount from the payments history.

This commit also adds in the purchase details screen a button to navigate
to the incoming payment that triggered the purchase.