(android) Add manual payments import/export (#685)
The user can now manually export the payments database to a public folder
of the device. The file is zipped and encrypted with ChaCha20Poly1305 using
the cloudkey already used by iOS.

When restoring a wallet, the user can also pick a payments backup file after
typing down the seed. This is optional.