(android) Support Lightning addresses in contacts (#699)
* Add sealed class grouping contacts payment codes

Bolt12 and LN addresses are similar as both are static payment instructions.
Representing that with a class allows us to factorise some code down the line.

* Refresh payments flow when contacts change

The payments flow need to be updated also when contacts change. To do
that we combine the payments flow from the payments-db with the contacts
list flow from the contacts-db.

This lets us remove workarounds in the iOS apps that are no longer needed.

* (android) Add UI for new contacts flow with LN address

Contacts can now contain Bolt12 codes or Lightning addresses (based
on LNURL or BIP353). The contact details screen can now handle the
creation or the editing of a contact, so we don't need a dedicated
component to create new contacts. The UI has also been streamlined.

---------

Co-authored-by: Robbie Hanson <304604+robbiehanson@users.noreply.github.com>