(android) Rework the receive screen for stateless offers
Phoenix can now generate Bolt12 offers at will, and give them
an amount and a description. This make them similar to Bolt11
invoices, and the UI can be simplified. The user can now press
a button that switches the Lightning payment request between
Bolt11 and Bolt12 seamlessly.

Note that stateless offers are not persisted in the database
and can still be paid even after reinstalling the app. The
default offer used previously is not used anymore, except
when claiming a BIP353 address.

Previous screens dedicated to Bolt12 have been removed.

This commit also upgrades Compose to 1.8 and targets API 35. It
makes use of the edge-to-edge design.