(android) Use the NFC reader mode to read NFC-A tags
To interact with an HCE device we need to use the NFC reader mode
with the ReaderCallback method, and implement the apdu commands
interaction with the tag, just like is done in the HCE service.

Reading a tag requires the user to tap a button in the Send screen.

This supersedes the changes made in fc8178eba.