(android) Use NFC to read/share payment requests (#722)
NFC is an alternative to scanning QR codes for sharing payment
requests between devices. The app needs to be able to read tags
to send, and emulate tags to receive.

* Added a service to emulate a NDEF tag

This service is manually started from the Receive screen. It
emits a NDEF text record containing a payment request (LN or
on-chain). When the service is started, a blocking dialog is
displayed.

* Use the NFC reader mode to read NFC-A tags

To interact with an HCE device we need to use the NFC reader mode
with the ReaderCallback method, and implement the apdu commands
interaction with the tag, just like is done in the HCE service.

To read a tag the user must tap a button in the Send screen.