(android) Support multiple fiat currencies in amount inputs
There is now a primary Bitcoin unit and a primary Fiat currency that
define how amounts are displayed throughout the app, just like before.

However inputs now accept an amount in another fiat or bitcoin unit.
It means that if the user is using Euro usually, but travels to the US,
he will be able to pick the USD currency when sending a payment, or when
generating an invoice.

A currency converter has been added to facilitate converting amounts to
other currencies. This converter is how the user can add new currencies
in the unit dropdowns.