(android) Support multiple fiat currencies (#726)
Inputs now accept an amount in another fiat or bitcoin unit than the 
primary options. It means that if the user is using Euro usually and want
to keep it that way, but travels to the US, he will be able to pick the USD
currency when sending a payment, or when generating an invoice.

A currency converter has been added to facilitate converting amounts to
other currencies. This converter is how the user can add new currencies
in the units dropdown.