(android) Prepare multiple seed system and rework business startup
This commit prepares the app for supporting multiple wallets. The seed
file can now store multiple mnemonics. The app is also not tied to a
single business object anymore, it can host and track several at the
same time.

Due to this change, the mechanism to start the business has been reworked.
This also fixes long standing issues coming from the legacy app where the
business was somehow tied to the foreground service and caused problems with
the background workers. We now use a singleton that can be accessed safely
from anywhere including the workers.

This commit is a WIP.