(android) Fix issues with double/string decimal conversion
Depending on the device locale, the decimal separator may be a comma
or a dot character. If it's a comma it causes problems when converting
the string input to a Double, because the method expects a dot.

This commit also prevents entering more than 2 decimals for fiat amounts
in the currency converter. It caused a problem where the additional 3rd
decimal (which is actually not visible) would cause the underlying amount
to be updated, which in turn refreshes the fiat amount ; it makes the UX
look very weird.