(android) Fix issues with device configuration changes
Removed scope from view models, and removed ambients shorthands that
assumed non-nullability too eagerly (e.g. business). This was causing
crashes when the device was rotated, or theme changed, because the
app was expecting values to exist in some screens, but they were not
yet restored.

Also, added a dialog to setup connection options when creating or
restoring a wallet.