(android) Update logic for wallet startup behaviour
A wallet should be automatically picked and started in some cases
which do not only depends on whether there's only 1 single wallet,
or if there's a default wallet. It also depends on how the app
got into the "locked" state.

For example, if the scren/device goes to sleep, the wallet should be
automatically locked, and as soon as the screen is on again, the
wallet that was used before should be loaded again, as soon as possible.