Move CurrencyManager outside business scope
The PhoenixBusiness object is tied to a specific wallet. It can be
destroyed, or many instances can be created. It's not suited for
routines that have application-wide scope, such as the Currency
Manager. Note that this is also true for a few other things, like
the network state monitoring, or the mempool.space fee estimation.

As such a PhoenixGlobal class has been added with the expecation
that it's created at application startup, and that it is unique.
Note that it probably could be a singleton, TBD.

The CurrencyManager logic is the same as before, though there are
some changes: the APIs objects have been moved to separate file in
a distinct package for readability. Also, the flow of watched
currencies is now stored in this Manager, and not in the
AppConfigManager.