(android) Support for multiple wallets (#742)
The user is now able to add, delete, and in general manage multiple
wallets on the same instance of Phoenix. Each wallet can have its own
name and avatar. Internally, wallets are indexed by the hash160 of the
wallet's public node id.

The seed file now stores multiple mnemonics.

The PhoenixBusiness object managing the actual node and coroutines is
now stored by a BusinessManager singleton. Data can be accessed safely
from anywhere including the background service workers. Several
businesses can run concurrently, although this can affect performance,
which will be fixed in a later commit.

Preferences are now per wallet and are managed by a DataStoreManager.
A new global preferences datastore has been added for preferences that
are global to the application, and not a given wallet.

Several ambients assuming non-nullable objects have been removed.

The startup and screenlock mechanism have been reworked to handle
multiple wallets. Specifically, a wallet selector screen lets the user picks
a wallet when needed, though there's also a system to load a wallet
automatically when relevant.

The active wallet is displayed in the Settings menu screen, as well as 
the Receive and Send screen (with incoming liquidity or balance for
convenience)

The navigation graph has been reworked and split into several files for
improved readability.

* (ios) Removed obsolete CloseDatabase parameter when stopping business

This parameter was introduced and used only by the Android app because of
long standing issues with the background system, which have been fixed with
the multiple wallets support update.

---------

Co-authored-by: Robbie Hanson <304604+robbiehanson@users.noreply.github.com>