Share managers between multiple nodes (#748)
The PhoenixBusiness object is tied to a specific wallet. It can be
destroyed, or many instances can be created. It's not suited for
routines that have application-wide scopes, such as the Currency
Manager.

A PhoenixGlobal class has been created with the expectation
that it's created at application startup, and that it is unique.

The CurrencyManager logic is the same as before, though there
has been some refactoring.

---------

Co-authored-by: Robbie Hanson <304604+robbiehanson@users.noreply.github.com>