(android) support fiat offers
Bolt12 offers may specify an amount, and that amount may be in a fiat
currency. The wallet should be able to understand that, and convert the
amount to btc, using the feerate fetched by the CurrencyManager.

A new method has been added in CurrencyManager to fetch the feearate of
a given fiat currency immediately (or use the cached value is cache is
fresh).

Since the conversion is not accurate, the user can change the amount to
send, even use an amount below what Phoenix computed. The recipient may
then reject the payment if it's too far from what it actually expects
for the given fiat amount.