(android) Disable legacy-app-status check at startup
Also always display the button to switch to the legacy app.

This commit allows user to be able to load the legacy app at
will, and attempt to restore very old channels data.

Note, the legacy-channel-status check has been removed from
the LSP and should not be used anymore, instead the switch is
done manually.